# Responsive-Web-Builder/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlutenFree Map API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /establishments/{id}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for an establishment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for an establishment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update own review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewBody"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete own review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Products
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Review:
      type: object
      properties:
        id:
          type: integer
        establishmentId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatarUrl:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        crossContaminationAlert:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - establishmentId
        - userId
        - userName
        - rating
        - crossContaminationAlert
        - createdAt
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - reviews
        - total
        - page
        - limit
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        crossContaminationAlert:
          type: boolean
        photoUrl:
          type: ["string", "null"]
      required:
        - rating
    UpdateReviewBody:
      type: object
      properties:
        rating:
          type: ["integer", "null"]
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        crossContaminationAlert:
          type: ["boolean", "null"]
        photoUrl:
          type: ["string", "null"]

    # Products
...